+++ examples/react/pages/index.tsx
@@ -25,7 +25,7 @@ import { setupEthereumWallets } from "@near-wallet-selector/ethereum-wallets";
 import { setupCoin98Wallet } from "@near-wallet-selector/coin98-wallet";
 import { setupIntearWallet } from "@near-wallet-selector/intear-wallet";
 import { setupSafePalWallet } from "@near-wallet-selector/safepal-wallet";
-
+import { setupWebAuthnWallet } from "@near-wallet-selector/webauthn-wallet";
 import { CONTRACT_ID, NETWORK_ID } from "../constants";
 import type { SetupParams } from "@near-wallet-selector/react-hook";
 
@@ -67,6 +67,9 @@ const walletSelectorConfig: SetupParams = {
     setupNearMobileWallet(),
     setupIntearWallet(),
     setupSafePalWallet(),
+    setupWebAuthnWallet({
+      relayerUrl: "http://localhost:7090/create-account", // Replace with your server's URL, example server is at packages/webauthn-wallet/relayer-examples/index.js
+    }),
   ],
 };