+++ packages/bitte-wallet/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @near-wallet-selector/bitte-wallet
 
+## 10.1.1
+
+## 10.1.0
+
 ## 10.0.0
 
 ### Minor Changes